Fixed bug in how the attribute operator comboBox is handled when the user selects an Attribute of type Date/Time.  e.g. startTime or endTime.
Added code to handle automatic setting of focus to DateTimePicker widget when user selects an attribute operator.  (I think the DateTimePicker might have a bug in it that causes it not to handle focus traversal properly or work with the FocusManager properly.  Perhaps send a bug report to lavantech.com and see if they will fix it.)